Convert all the HWRPs to Chef custom resources
