+++ libraries/matchers.rb
@@ -21,12 +21,10 @@
 
 if defined?(ChefSpec)
   {
-    dnsmasq_local: [:create, :remove],
-    dnsmasq_local_app: [:install, :remove],
-    dnsmasq_local_config: [:create, :remove],
-    dnsmasq_local_service: [
-      :create, :remove, :enable, :disable, :stop, :start, :restart
-    ]
+    dnsmasq_local: %i[create remove],
+    dnsmasq_local_app: %i[install remove],
+    dnsmasq_local_config: %i[create remove],
+    dnsmasq_local_service: %i[create remove enable disable stop start restart]
   }.each do |matcher, actions|
     ChefSpec.define_matcher(matcher)
