+++ libraries/resource_dnsmasq_local_service_rhel_sysvinit.rb
@@ -1,8 +1,9 @@
 # encoding: utf-8
 # frozen_string_literal: true
+
 #
 # Cookbook Name:: dnsmasq-local
-# Library:: provider_dnsmasq_local_service_rhel_sysvinit
+# Library:: resource_dnsmasq_local_service_rhel_sysvinit
 #
 # Copyright 2016, Socrata, Inc.
 #
@@ -19,19 +20,17 @@
 # limitations under the License.
 #
 
-require_relative 'provider_dnsmasq_local_service_rhel'
+require_relative 'resource_dnsmasq_local_service_rhel'
 
 class Chef
-  class Provider
-    # A Chef provider for Sysvinit RHEL services.
+  class Resource
+    # A Chef resource for Sysvinit RHEL services.
     #
     # @author Jonathan Hartman <jonathan.hartman@socrata.com>
     class DnsmasqLocalServiceRhelSysvinit < DnsmasqLocalServiceRhel
-      if defined?(provides)
-        provides :dnsmasq_local_service,
-                 platform_family: 'rhel',
-                 platform_version: '< 7'
-      end
+      provides :dnsmasq_local_service,
+               platform_family: 'rhel',
+               platform_version: '< 7'
 
       #
       # Patch the RHEL < 7 init script to pass on the DNSMASQ_OPTS environment
